pyproject.toml
[tool.poetry]
name = "Backend"
version = "0.1.0"
description = ""
authors = ["MelianMiko <support@mmk.pw>"]
package-mode = false

[tool.poetry.dependencies]
python = "<3.13,>=3.12"
fastapi = "^0.112.1"
tensorflow = "^2.18.0"
pillow = "^11.0.0"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

